backend/requirements.txt
pandas
pybaseball
flask-sqlalchemy
psycopg2-binary
python-dotenv